Backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-auth
requests
python-jose
web3